confirm to delete

+++ js/addBook.js
@@ -39,15 +39,18 @@ async function addBook(event) {
 }
 
 async function deleteBook(id) {
-  const response = await fetch(
-    `https://elibraryrestapi.herokuapp.com/elibrary/api/book/delete/${id}`,
-    {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    }
-  );
-  alert("Book was deleted!");
-  location.reload();
+  let result = confirm("This book will be removed. Do you want to continue?");
+  if (result) {
+    const response = await fetch(
+      `https://elibraryrestapi.herokuapp.com/elibrary/api/book/delete/${id}`,
+      {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      }
+    );
+    alert("Book was deleted!");
+    location.reload();
+  }
 }
